requirements.txt
# Database
psycopg2-binary>=2.9.9
PyYAML>=6.0.1

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.4.0

# Visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

